make app, tier and node id as long + unit test for appdynamics apis

+++ rest/src/main/java/software/wings/resources/AppdynamicsResource.java
@@ -17,7 +17,6 @@ import software.wings.service.impl.newrelic.NewRelicApplication;
 import software.wings.service.impl.newrelic.NewRelicMetricAnalysisRecord;
 import software.wings.service.intfc.MetricDataAnalysisService;
 import software.wings.service.intfc.appdynamics.AppdynamicsService;
-import software.wings.service.intfc.newrelic.NewRelicService;
 import software.wings.sm.StateType;
 
 import java.io.IOException;
@@ -53,7 +52,7 @@ public class AppdynamicsResource {
   @Timed
   @ExceptionMetered
   public RestResponse<List<AppdynamicsTier>> getAllTiers(@QueryParam("accountId") String accountId,
-      @QueryParam("settingId") final String settingId, @QueryParam("appdynamicsAppId") int appdynamicsAppId)
+      @QueryParam("settingId") final String settingId, @QueryParam("appdynamicsAppId") long appdynamicsAppId)
       throws IOException {
     return new RestResponse<>(appdynamicsService.getTiers(settingId, appdynamicsAppId));
   }
@@ -63,8 +62,8 @@ public class AppdynamicsResource {
   @Timed
   @ExceptionMetered
   public RestResponse<List<AppdynamicsNode>> getAllNodes(@QueryParam("accountId") String accountId,
-      @QueryParam("settingId") final String settingId, @QueryParam("appdynamicsAppId") int appdynamicsAppId,
-      @QueryParam("tierId") int tierId) throws IOException {
+      @QueryParam("settingId") final String settingId, @QueryParam("appdynamicsAppId") long appdynamicsAppId,
+      @QueryParam("tierId") long tierId) throws IOException {
     return new RestResponse<>(appdynamicsService.getNodes(settingId, appdynamicsAppId, tierId));
   }
 
@@ -74,7 +73,7 @@ public class AppdynamicsResource {
   @ExceptionMetered
   public RestResponse<List<AppdynamicsBusinessTransaction>> getAllBusinessTransactions(
       @QueryParam("accountId") String accountId, @QueryParam("settingId") final String settingId,
-      @QueryParam("appdynamicsAppId") int appdynamicsAppId) throws IOException {
+      @QueryParam("appdynamicsAppId") long appdynamicsAppId) throws IOException {
     return new RestResponse<>(appdynamicsService.getBusinessTransactions(settingId, appdynamicsAppId));
   }
 
@@ -83,7 +82,7 @@ public class AppdynamicsResource {
   @Timed
   @ExceptionMetered
   public RestResponse<List<AppdynamicsMetric>> getTierBTMetrics(@QueryParam("settingId") final String settingId,
-      @QueryParam("appdynamicsAppId") int appdynamicsAppId, @QueryParam("tierId") int tierId)
+      @QueryParam("appdynamicsAppId") long appdynamicsAppId, @QueryParam("tierId") long tierId)
       throws IOException, InterruptedException {
     return new RestResponse<>(appdynamicsService.getTierBTMetrics(settingId, appdynamicsAppId, tierId));
   }
@@ -93,7 +92,7 @@ public class AppdynamicsResource {
   @Timed
   @ExceptionMetered
   public RestResponse<List<AppdynamicsMetricData>> getTierBTMetricData(@QueryParam("settingId") final String settingId,
-      @QueryParam("appdynamicsAppId") int appdynamicsAppId, @QueryParam("tierId") int tierId,
+      @QueryParam("appdynamicsAppId") long appdynamicsAppId, @QueryParam("tierId") long tierId,
       @QueryParam("btName") String btName, @QueryParam("duration-in-mins") int durationInMinutes) throws IOException {
     return new RestResponse<>(
         appdynamicsService.getTierBTMetricData(settingId, appdynamicsAppId, tierId, btName, durationInMinutes));
