+++ rest/src/main/java/software/wings/service/intfc/appdynamics/AppdynamicsService.java
@@ -19,9 +19,9 @@ import javax.validation.constraints.NotNull;
 public interface AppdynamicsService {
   List<NewRelicApplication> getApplications(@NotNull String settingId) throws IOException;
 
-  List<AppdynamicsTier> getTiers(String settingId, int appdynamicsAppId) throws IOException;
+  List<AppdynamicsTier> getTiers(String settingId, long appdynamicsAppId) throws IOException;
 
-  List<AppdynamicsNode> getNodes(String settingId, int appdynamicsAppId, int tierId) throws IOException;
+  List<AppdynamicsNode> getNodes(String settingId, long appdynamicsAppId, long tierId) throws IOException;
 
   List<AppdynamicsBusinessTransaction> getBusinessTransactions(@NotNull String settingId, @Valid long appdynamicsAppId)
       throws IOException;
@@ -31,6 +31,6 @@ public interface AppdynamicsService {
   List<AppdynamicsMetric> getTierBTMetrics(@NotNull String settingId, long appdynamicsAppId, long tierId)
       throws IOException;
 
-  List<AppdynamicsMetricData> getTierBTMetricData(@NotNull String settingId, int appdynamicsAppId, int tierId,
+  List<AppdynamicsMetricData> getTierBTMetricData(@NotNull String settingId, long appdynamicsAppId, long tierId,
       @NotNull String btName, int durantionInMinutes) throws IOException;
 }
